+++ test/fast_localization_test.dart
@@ -47,4 +47,27 @@ void main() {
     expect(enName, 'Hasan Al-Yazidi');
     expect(arName, 'حسن اليزيدي');
   });
+
+  test('Localization.translate() aliasess', () async {
+    SharedPreferences.setMockInitialValues({});
+
+    final locales = {
+      Locale('en'): {
+        "info": {
+          "id": "20200209",
+          "name": "My name is {{name}}",
+          "age": "I am {{age}} years old",
+        },
+      },
+    };
+
+    await Localization.load(locales);
+
+    expect(Localization.t('info.id'), "20200209");
+    expect(Localization.t('info.name', {"name": "Hasan"}), 'My name is Hasan');
+    expect(Localization.t('info.age', {"age": "128"}), 'I am 128 years old');
+    expect(t('info.id'), "20200209");
+    expect(t('info.name', {"name": "Hasan"}), 'My name is Hasan');
+    expect(t('info.age', {"age": "128"}), 'I am 128 years old');
+  });
 }